+++ packages/jest-config/src/ValidConfig.ts
@@ -103,7 +103,10 @@ const initialOptions: Config.InitialOptions = {
   snapshotResolver: '<rootDir>/snapshotResolver.js',
   snapshotSerializers: ['my-serializer-module'],
   testEnvironment: 'jest-environment-jsdom',
-  testEnvironmentOptions: {userAgent: 'Agent/007'},
+  testEnvironmentOptions: {
+    url: 'http://localhost',
+    userAgent: 'Agent/007',
+  },
   testFailureExitCode: 1,
   testLocationInResults: false,
   testMatch: ['**/__tests__/**/*.[jt]s?(x)', '**/?(*.)+(spec|test).[jt]s?(x)'],
@@ -117,7 +120,6 @@ const initialOptions: Config.InitialOptions = {
   testRunner: 'jasmine2',
   testSequencer: '@jest/test-sequencer',
   testTimeout: 5000,
-  testURL: 'http://localhost',
   timers: 'real',
   transform: {
     '\\.js$': '<rootDir>/preprocessor.js',