+++ packages/jest-types/src/Config.ts
@@ -90,7 +90,7 @@ export type DefaultOptions = {
   slowTestThreshold: number;
   snapshotSerializers: Array<Path>;
   testEnvironment: string;
-  testEnvironmentOptions: Record<string, unknown>;
+  testEnvironmentOptions?: Record<string, any>;
   testFailureExitCode: string | number;
   testLocationInResults: boolean;
   testMatch: Array<Glob>;
@@ -98,7 +98,6 @@ export type DefaultOptions = {
   testRegex: Array<string>;
   testRunner: string;
   testSequencer: string;
-  testURL: string;
   timers: Timers;
   transformIgnorePatterns: Array<Glob>;
   useStderr: boolean;
@@ -201,7 +200,7 @@ export type InitialOptions = Partial<{
   snapshotSerializers: Array<Path>;
   errorOnDeprecated: boolean;
   testEnvironment: string;
-  testEnvironmentOptions: Record<string, unknown>;
+  testEnvironmentOptions?: Record<string, any>;
   testFailureExitCode: string | number;
   testLocationInResults: boolean;
   testMatch: Array<Glob>;
@@ -212,7 +211,6 @@ export type InitialOptions = Partial<{
   testResultsProcessor: string;
   testRunner: string;
   testSequencer: string;
-  testURL: string;
   testTimeout: number;
   timers: Timers;
   transform: {
@@ -363,13 +361,12 @@ export type ProjectConfig = {
   snapshotResolver?: Path;
   snapshotSerializers: Array<Path>;
   testEnvironment: string;
-  testEnvironmentOptions: Record<string, unknown>;
+  testEnvironmentOptions?: Record<string, any>;
   testMatch: Array<Glob>;
   testLocationInResults: boolean;
   testPathIgnorePatterns: Array<string>;
   testRegex: Array<string | RegExp>;
   testRunner: string;
-  testURL: string;
   timers: Timers;
   transform: Array<[string, Path, Record<string, unknown>]>;
   transformIgnorePatterns: Array<Glob>;
@@ -452,7 +449,6 @@ export type Argv = Arguments<
     testResultsProcessor: string;
     testRunner: string;
     testSequencer: string;
-    testURL: string;
     testTimeout: number | null | undefined;
     timers: string;
     transform: string;